[ADD] src/utils2.c
